Kinda improved physics se

+++ SF_Cairo/DE.h
@@ -40,7 +40,7 @@ void Draw_Mine (cairo_t *cr, int x, int y, int size);
 void Draw_Missile (cairo_t *cr, int x, int y, int Headings, int size, int missile_idx);
 void Draw_Shell(cairo_t *cr, int x, int y, int Headings, int size);
 
-int Find_Headings(int x1,int y1,int x2,int y2);
+float Find_Headings(int x1,int y1,int x2,int y2);
 
 //void set_initial_vals(cairo_t *cr);
 void start_drawing();
